Enemy basics
Created two test enemies and let them move to the player. They now keep
tracking the player, need to make them simply move to the player when
they are created and then keep moving in that direction later (at least
for some enemies). Oh, and the code needs a cleanup too, but this is
okay for now.